requirements.txt
pip~=23.2.1
wheel~=0.41.2
Jinja2~=3.1.4
MarkupSafe~=2.1.5
Werkzeug~=3.0.3
click~=8.1.7
blinker~=1.8.2
itsdangerous~=2.2.0
setuptools~=68.2.0
certifi~=2024.7.4
urllib3~=2.2.2
idna~=3.7
Flask~=3.0.3
requests~=2.32.3
python-dotenv~=1.0.1